moved validations into validation file under utils and removed commented unused codes
