+++ backend/db/seeders/20230415035250-spot-seeder.js
@@ -5,7 +5,6 @@ if (process.env.NODE_ENV === 'production') {
   options.schema = process.env.SCHEMA;  // define your schema in options object
 }
 
-// const {Spot} = require("../models");
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
